dynamic_alerts/requirements.txt
opencv-python
psycopg2-binary
python-dotenv
requests
flask